+++ lib/src/xml_events/converters/node_decoder.dart
@@ -78,8 +78,8 @@ class _XmlNodeDecoderSink extends ChunkedConversionSink<List<XmlEvent>>
       commit(XmlDeclaration(convertAttributes(event.attributes)), event);
 
   @override
-  void visitDoctypeEvent(XmlDoctypeEvent event) =>
-      commit(XmlDoctype(event.text), event);
+  void visitDoctypeEvent(XmlDoctypeEvent event) => commit(
+      XmlDoctype(event.name, event.externalId, event.internalSubset), event);
 
   @override
   void visitEndElementEvent(XmlEndElementEvent event) {